% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_geo_utils.R
\name{calc_area_weighted_values}
\alias{calc_area_weighted_values}
\title{Rasterize polygons - calculate area weighted values per cell}
\usage{
calc_area_weighted_values(rast_obj, poly_vect_obj, layer)
}
\arguments{
\item{rast_obj}{terra raster object, e.g. output of create_berlin_rast()}

\item{poly_vect_obj}{terra SpatVector object with geometry polygons}

\item{layer}{layer name that contains the values to be rasterized}
}
\value{
terra SpatRaster object in crs ESPG:25833
}
\description{
raster_obj and poly_vect_obj need to be in the same crs, preferibly with a
crs that has meteres as units.
}
