% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_streets.R
\name{calc_line_length_per_cell}
\alias{calc_line_length_per_cell}
\title{Create terra SpatRaster object with line lengths per cell}
\usage{
calc_line_length_per_cell(rast_obj, lines_vect_obj)
}
\arguments{
\item{rast_obj}{terra raster object, e.g. output of create_berlin_rast()}

\item{lines_vect_obj}{terra SpatVector object with geometry lines}
}
\value{
terra SpatRaster object in crs ESPG:25833
}
\description{
raster_obj and lines_obj need to be in the same crs, preferibly with a
crs that has meteres as units.
}
