% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_traffic_intensity.R
\name{calc_traffic_intensity}
\alias{calc_traffic_intensity}
\title{Calculates scaled traffic intensity per raster cell.}
\usage{
calc_traffic_intensity(vehicle_type, street_lengths)
}
\arguments{
\item{vehicle_type}{character, either "kfz" or "lkw"}

\item{street_lengths}{data.frame, e.g. output from "traffic_intensity" query
where each row represents one raster cell}
}
\value{
numeric vector, traffic intensity for vehicle type scaled
}
\description{
Scaling factor is the Berlin overall average of unscaled traffic intensity.
Interpretation of traffic intensity is:
1 = traffic equals that of an average cell which has streets in it
smaller than 1 = less traffic than an average cell which has streets in it
bigger than 1 = more traffic than an average cell which has streets in it
}
