% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_traffic_volume.R
\name{calc_traffic_volume}
\alias{calc_traffic_volume}
\title{Calculates rasterized traffic volume per 24h.}
\usage{
calc_traffic_volume(traffic_volume_exact, traffic_volume_avg, berlin_rast)
}
\arguments{
\item{traffic_volume_exact}{SpatVector (lines), exact traffic volumes}

\item{traffic_volume_avg}{SpatVector (lines), average traffic volumes}

\item{berlin_rast}{SpatRaster of Berlin}
}
\value{
SpatRaster with one layer representing the "gap" filled max of
traffic volume per 24h
}
\description{
This function takes the exact traffic volume as SpatVector (lines) and
rasterizes the values onto berlin_rast via the max value found in each cell.
For cells where there is no valid exact traffic volume, this function fills
these "gaps" with the average traffic volume (SpatVector lines) if there are
any in this cell. Returns a SpatRaster with "gap filled" traffic volume.
}
