% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_trees.R
\name{calc_tree_features}
\alias{calc_tree_features}
\title{Create terra SpatRaster object tree count, mean age and mean height per cell}
\usage{
calc_tree_features(trees_vect, berlin_rast)
}
\arguments{
\item{trees_vect}{terra SpatVector object with trees as points and
"standalter" and "baumhoehe" as numeric layers}

\item{berlin_rast}{terra raster object, e.g. output of create_berlin_rast()}
}
\value{
terra SpatRaster object with layers "count", "age" and "height".
}
\description{
Inputs trees_vect and berlin_rast need to be in the same crs,
preferably ESPG:25833.
}
