% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_berlin_raster.R
\name{create_berlin_rast}
\alias{create_berlin_rast}
\title{Create Berlin terra "SpatRaster" object}
\usage{
create_berlin_rast(cell_size = 50)
}
\arguments{
\item{cell_size}{(numeric) How wide and high a cell of the
resulting "SpatRaster" should be? Measured in metres.
Defaults to 50m.}
}
\value{
SpatRaster object in crs ESPG:25833
}
\description{
Create Berlin terra "SpatRaster" object
}
