% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_berlin_raster.R
\name{create_berlin_stars}
\alias{create_berlin_stars}
\title{Create stars raster for Berlin}
\usage{
create_berlin_stars(cell_size = 50, target_crs = 25833)
}
\arguments{
\item{cell_size}{(numeric) How wide and high (horizontal and vertical extension)
a cell of the resulting stars raster should be?
Measured in metres. Defaults to 50m.}

\item{target_crs}{(int) What crs (= Coordinate Reference System) should
the resulting stars raster have? Defaults to EPSG:25833
(https://epsg.io/25833), which constitutes the standard on Berlin's
Fisbroker Geoservers. Conveniently, EPSG:25833 uses metres as units.}
}
\value{
stars object - raster of berlin
}
\description{
This function creates the target raster (=grid) for Berlin.
Default values create a grid of 50m x 50m cells.
}
