% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_reprojection.R
\name{create_reprojection_mapping}
\alias{create_reprojection_mapping}
\title{Create reprojection mapping table in DB}
\usage{
create_reprojection_mapping()
}
\description{
Reprojection mapping table: table with distinct
lat and lon as integer (lat_int, lon_int)
mapped to x, y coordinates.
}
