% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_geo_utils.R
\name{df_to_rast}
\alias{df_to_rast}
\title{Transforms a data.frame into a terra SpatRaster object.}
\usage{
df_to_rast(df, crs = 25833, drop_id = TRUE)
}
\arguments{
\item{df}{(data.frame) As created by rast_to_df()}

\item{crs}{(integer) EPSG id for crs, defaults to 25833}

\item{drop_id}{(bool)}
}
\value{
(spatRaster) A terra spatRaster object .
}
\description{
Transforms a data.frame into a terra SpatRaster object.
}
