% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_traffic_utils.R
\name{exclude_non_drivable_streets}
\alias{exclude_non_drivable_streets}
\title{Exclude all streets of category V which are non car drivable.}
\usage{
exclude_non_drivable_streets(streets_raw_sf)
}
\arguments{
\item{streets_raw_sf}{sf object of raw streets (spatial lines) data}
}
\value{
sf object of streets. All streets in this sf object all car drivable.
Walking paths etc. were excluded.
}
\description{
More info on categories here:
https://de.wikipedia.org/wiki/\%C3\%9Cbergeordnetes_Stra\%C3\%9Fennetz_von_Berlin
"G" stands for Gemeindestraße, categories such as "F" (Fußgängerobjekte) get
excluded. This functtion reduces the number of streets from roughly 43000 to
37000.
}
