% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_coord_stadt_berlin.R
\name{extract.coord_stadt_berlin}
\alias{extract.coord_stadt_berlin}
\title{Extract for coord_stadt_berlin}
\usage{
\method{extract}{coord_stadt_berlin}(x, ...)
}
\arguments{
\item{x}{clickhouse single_source}

\item{...}{other parameters}
}
\description{
Infact it is an empty function - we need no extraction from database.
But if we don't have an extract function, etl does not work.
}
