% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_extract_raw_clickhouse.R
\name{extract_raw_clickhouse}
\alias{extract_raw_clickhouse}
\title{Gets raw geodata from clickhouse}
\usage{
extract_raw_clickhouse(query)
}
\arguments{
\item{query}{(character) name of sql file or a query as string.}
}
\value{
sf object
}
\description{
Downloads data from clickhouse and turns geometry column from geojson into sf geometry.
Queried data must contain a 'geometry' column, that contains geo information in the form of
geojson. Returns an sf object.
}
