% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_berlin_raster.R
\name{get_berlin_bezirke}
\alias{get_berlin_bezirke}
\title{Get the Berlin boroughs as an sf object (polygons)}
\usage{
get_berlin_bezirke()
}
\value{
sf object - 12 polygons of the Berlin boroughs
}
\description{
Get the Berlin boroughs as an sf object (polygons)
}
