% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_traffic_intensity.R
\name{get_overall_avg_traffic_intensity_unscaled}
\alias{get_overall_avg_traffic_intensity_unscaled}
\title{Get Berlin wide overall average of traffic intensity}
\usage{
get_overall_avg_traffic_intensity_unscaled(vehicle_type, street_lengths)
}
\arguments{
\item{vehicle_type}{character, either "kfz" or "lkw"}

\item{street_lengths}{data.frame, e.g. output from "traffic_intensity" query}
}
\value{
numeric scalar, overall average of traffic intensity unscaled
}
\description{
Get Berlin wide overall average of traffic intensity as unscaled scalar.
Traffic intensity unscaled is defined as product of street length in metres
multiplied with number of vehicles - all by street type.
Overall average is then the sum over all street types.
}
