% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_traffic_intensity.R
\name{get_street_length_avgs}
\alias{get_street_length_avgs}
\title{Get average street lengths per cell by strassenklasse}
\usage{
get_street_length_avgs(street_lengths)
}
\arguments{
\item{street_lengths}{data.frame, e.g. output from "traffic_intensity" query}
}
\value{
data.frame in long format with columns "strassenklasse" and "length"
}
\description{
Get average street lengths per cell by strassenklasse
}
