% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_traffic_utils.R
\name{get_traffic_avgs}
\alias{get_traffic_avgs}
\title{Get average traffic numbers per street type for a particular vehicle type}
\usage{
get_traffic_avgs(vehicle_type)
}
\arguments{
\item{vehicle_type}{character, either "kfz" or "lkw"}
}
\value{
data.frame in wide format with only one row for vehicle_type
}
\description{
Get average traffic numbers per street type for a particular vehicle type
}
