% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_traffic_volume.R
\name{get_traffic_volume_avg}
\alias{get_traffic_volume_avg}
\title{Get average traffic volume as sf object (streets as lines)}
\usage{
get_traffic_volume_avg(vehicle_type = "kfz")
}
\arguments{
\item{vehicle_type}{character, either "kfz" or "lkw"}
}
\value{
sf object (lines), all streets
}
\description{
Get average traffic volume as sf object (streets as lines)
}
