% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_buildings.R
\name{prepare_buildings_density}
\alias{prepare_buildings_density}
\title{Prepare raw building density data}
\usage{
prepare_buildings_density(buildings_density_raw, as_vect = TRUE)
}
\arguments{
\item{buildings_density_raw}{sf object of raw building density (polygon) data}

\item{as_vect}{bool, shall return value be a terra SpatVect object?
Defaults to TRUE. If false, an sf object is returned.}
}
\value{
SpatVect terra object in crs ESPG:25833, sf object if
as_vect = FALSE
}
\description{
Prepare raw building density data
}
