% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_land_use.R
\name{prepare_land_use}
\alias{prepare_land_use}
\title{Prepare raw land use data for coverage calculation and collapse use categories.}
\usage{
prepare_land_use(land_use_raw)
}
\arguments{
\item{land_use_raw}{sf object of raw land_use (polygons) data}
}
\value{
sf object of clean land_use data with collapsed factor levels for use
}
\description{
transform the land_use data
}
