% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_traffic_utils.R
\name{prepare_streets}
\alias{prepare_streets}
\title{Prepare raw streets data for various features}
\usage{
prepare_streets(streets_raw_sf, as_vect = TRUE)
}
\arguments{
\item{streets_raw_sf}{sf object of raw streets (spatial lines) data}

\item{as_vect}{bool, shall return value be a terra SpatVect object?
Defaults to TRUE. If false, an sf object is returned.}
}
\value{
SpatVect terra object in crs ESPG:25833, sf object if
as_vect = FALSE
}
\description{
Prepare raw streets data for various features
}
