% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_traffic_volume.R
\name{prepare_traffic}
\alias{prepare_traffic}
\title{Prepare raw exact traffic volume for rasterization}
\usage{
prepare_traffic(traffic_raw, vehicle_type = "kfz")
}
\arguments{
\item{traffic_raw}{sf object, "Verkehrsmengen" (=traffic volume) from raw db}

\item{vehicle_type}{character, either "kfz" or "lkw"}
}
\value{
sf object (lines), representing exact traffic volumes
}
\description{
Prepare raw exact traffic volume for rasterization
}
