% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_trees.R
\name{prepare_trees}
\alias{prepare_trees}
\title{Prepare raw tree data for calculation of tree count, mean age and mean height}
\usage{
prepare_trees(trees_raw)
}
\arguments{
\item{trees_raw}{sf object of raw park and street trees (spatial points) data -
must include columns "standalter" and "baumhoehe"}
}
\value{
sf object excluding inplausible outliers in "baumhoehe" and "standalter",
however NAs for these attributes are kept. Crs is 25833.
}
\description{
transform the trees data
}
\details{
Excludes outliers and reprojects trees data to crs 25833.
}
