% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_geo_utils.R
\name{rast_to_df}
\alias{rast_to_df}
\title{Turns a terra SpatRaster Object into a data.frame.}
\usage{
rast_to_df(spat_raster_obj, add_id = TRUE, na_rm = TRUE)
}
\arguments{
\item{spat_raster_obj}{(spatRaster) a spatRaster object (terra package).}

\item{add_id}{(logical) should a running id be added to data.frame as the
first column?}

\item{na_rm}{(logical) If TRUE, cells that have a NA value in at least one
layer are removed. Defaults to TRUE.}
}
\value{
data.frame with coordinate columns x and y. These represent the
centroids of the raster cells. Values of the SpatRaster layers are
represented in the values column. No crs info in this return object!
}
\description{
Turns a terra SpatRaster Object into a data.frame while loosing the crs
(= coordinate reference system) info is lost. One former raster cell is
represented as one row with centroid coordinates.
}
