% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_traffic_volume.R
\name{rasterize_lines_max}
\alias{rasterize_lines_max}
\title{Rasterizes lines by assigning maximum value of attribute to cell}
\usage{
rasterize_lines_max(lines_sf_obj, rast_obj, layer_name = "kfz", n = 1)
}
\arguments{
\item{lines_sf_obj}{sf object (lines)}

\item{rast_obj}{SpatRaster}

\item{layer_name}{name of layer in lines_vect_obj to rasterized,
defaults to "kfz"}

\item{n}{numeric, number of points to be sampled from lines per metre before
rasterization}
}
\value{
SpatRaster with one layer representing max of values
}
\description{
Rasterizes lines by assigning maximum value of attribute to cell
}
