% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_reprojection.R
\name{reproject}
\alias{reproject}
\title{Reprojection of lat/lon to x/y.}
\usage{
reproject(df)
}
\arguments{
\item{df}{data.frame with lat, lon}
}
\value{
data.frame
}
\description{
lat, lon (EPSG:4326/WGS84) -> x, y (EPSG:25833/ETRS89)
}
\details{
Convert df to spatial object and
transform lat/lon to x/y.
}
