% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_coord_mapping_stadt_streets.R
\name{round_line_coordinates}
\alias{round_line_coordinates}
\title{Rounds coordinates of linestring sf object}
\usage{
round_line_coordinates(lines_sf, decimals = 2)
}
\arguments{
\item{lines_sf}{spatial sf object of lines (Linestring) in crs 25833}

\item{decimals}{integer - number of decimals to keep. Defaults to 2 decimals.
Precision: 1 decimal = 10cm, 2 decimals = 1cm.}
}
\value{
spatial sf object of lines with rounded geometry column
}
\description{
Rounds coordinates of linestring sf object
}
