% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_send_data_clickhouse.R
\name{send_data_clickhouse}
\alias{send_data_clickhouse}
\title{Send data wrapper}
\usage{
send_data_clickhouse(df, table, mode = "replace", ...)
}
\arguments{
\item{df}{(data.frame) a data frame with the data.}

\item{table}{(str) name of the table in the database.}

\item{mode}{(str) "replace" (default), "insert" or "truncate".}

\item{...}{arbitrary arguments for parameterized queries.
For "replace", the data is sent with "insert" and then the table is optimized to remove
duplicates w.r.t. the order statement (only if the sent data has more than 0 rows).
For "truncate", the table is first truncated (even if the new data is empty).}
}
\description{
This is a wrapper around \link[dbtools]{sendData} to easily reference SQL files and
use credentials handled inside the package.
}
