% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_geo_utils.R
\name{sf_obj_check}
\alias{sf_obj_check}
\title{Checks if obj is a spatial sf object.}
\usage{
sf_obj_check(obj)
}
\arguments{
\item{obj}{any object}
}
\description{
Passes if obj is a spatial sf object, throws error otherwise.
}
