% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_geo_utils.R
\name{sf_to_geojson_df}
\alias{sf_to_geojson_df}
\title{Turns sf object into a dataframe}
\usage{
sf_to_geojson_df(sf_obj)
}
\arguments{
\item{sf_obj}{spatial sf object}
}
\value{
dataframe with a geometry column that is of type character (geojson)
}
\description{
Turns the geometry of a sf_obj into a character column, in the form of a geojson.
}
