% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_sources.R, R/01_single_sources.R
\name{data_sources}
\alias{data_sources}
\alias{single_source}
\title{Data sources to query from clickhouse}
\usage{
data_sources()

single_source(
  name,
  get_data = extract_raw_clickhouse,
  load_mode = "replace",
  send_data = send_data_clickhouse,
  ...
)
}
\arguments{
\item{name}{(character) a unique name used as class and to identify target
table on database.}

\item{get_data}{(function)}

\item{load_mode}{(character) defaults to "replace"}

\item{send_data}{(function)}

\item{...}{arguments used within methods and available for queries}
}
\description{
List data sources. This list defines which data source to extract from clickhouse.
}
