% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_buildings.R
\name{transform.buildings}
\alias{transform.buildings}
\title{Transform for buildings}
\usage{
\method{transform}{buildings}(x, cell_size = 50, ...)
}
\arguments{
\item{x}{clickhouse single_source}

\item{cell_size}{cell_size of regular target rast, measured in metres}

\item{...}{other parameters}
}
\description{
Transform for buildings
}
