% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_coord_mapping_stadt_reprojection.R
\name{transform.coord_mapping_stadt_reprojection}
\alias{transform.coord_mapping_stadt_reprojection}
\title{Transform for coord_mapping_stadt_reprojection}
\usage{
\method{transform}{coord_mapping_stadt_reprojection}(x, ...)
}
\arguments{
\item{x}{clickhouse single_source}

\item{...}{other parameters}
}
\description{
Transform for coord_mapping_stadt_reprojection
}
