% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_coord_mapping_stadt_streets.R
\name{transform.coord_mapping_stadt_streets}
\alias{transform.coord_mapping_stadt_streets}
\title{Transform for coord_mapping_stadt_streets}
\usage{
\method{transform}{coord_mapping_stadt_streets}(x, cell_size = 50, ...)
}
\arguments{
\item{x}{clickhouse single_source}

\item{cell_size}{cell_size of regular target rast, measured in metres}

\item{...}{other parameters}
}
\description{
Transform for coord_mapping_stadt_streets
}
