% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_coord_stadt_berlin.R
\name{transform.coord_stadt_berlin}
\alias{transform.coord_stadt_berlin}
\title{Transform for coord_stadt_berlin}
\usage{
\method{transform}{coord_stadt_berlin}(x, cell_size = 50, ...)
}
\arguments{
\item{x}{clickhouse single_source}

\item{cell_size}{cell_size of regular target rast, measured in metres}

\item{...}{other parameters}
}
\description{
Transform for coord_stadt_berlin
}
