% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_land_use.R
\name{transform.land_use}
\alias{transform.land_use}
\title{Transform for land_use}
\usage{
\method{transform}{land_use}(x, cell_size = 50, ...)
}
\arguments{
\item{x}{clickhouse single_source}

\item{cell_size}{cell_size of regular target rast, measured in metres}

\item{...}{other parameters}
}
\description{
Transform for land_use
}
