% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_streets.R
\name{transform.streets}
\alias{transform.streets}
\title{Transform for network streets}
\usage{
\method{transform}{streets}(x, cell_size = 50, ...)
}
\arguments{
\item{x}{clickhouse single_source}

\item{cell_size}{cell_size of regular target rast, measured in metres}

\item{...}{other parameters}
}
\description{
Transform for network streets
}
