% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_traffic_intensity.R
\name{transform.traffic_intensity}
\alias{transform.traffic_intensity}
\title{Transform for traffic intensity}
\usage{
\method{transform}{traffic_intensity}(x, ...)
}
\arguments{
\item{x}{clickhouse single_source}

\item{...}{other parameters}
}
\description{
Transform for traffic intensity
}
