% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_traffic_volume.R
\name{transform.traffic_volume}
\alias{transform.traffic_volume}
\title{Transform for traffic_volume}
\usage{
\method{transform}{traffic_volume}(x, cell_size = 50, ...)
}
\arguments{
\item{x}{clickhouse single_source}

\item{cell_size}{cell_size of regular target rast, measured in metres}

\item{...}{other parameters}
}
\description{
Transform for traffic_volume
}
