% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_trees.R
\name{transform.trees}
\alias{transform.trees}
\title{Transform for trees}
\usage{
\method{transform}{trees}(x, cell_size = 50, ...)
}
\arguments{
\item{x}{clickhouse single_source}

\item{cell_size}{cell_size of regular target raster, measured in metres}

\item{...}{other parameters}
}
\description{
Transform for trees
}
